Fix one more warning
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@24841 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/rcfpch/rcf_pch.c
@@ -1044,7 +1044,7 @@ rcf_pch_run(const char *confstr, const char *info)
                         }
                         else
                         {
-                            SEND_ANSWER("0 %d", pid);
+                            SEND_ANSWER("0 %ld", (long)pid);
                         }
                         
                         break;
